requirements.txt
accelerate==0.20.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
clip-interrogator==0.6.0
colorama==0.4.6
filelock==3.12.0
Flask==2.3.2
fsspec==2023.5.0
ftfy==6.1.1
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
open-clip-torch==2.20.0
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.5
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
sklearn==0.0.post5
sympy==1.12
threadpoolctl==3.1.0
timm==0.9.2
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.6.3
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.3.4
zipp==3.15.0